+++ WTF/Account/197404344#2/SavedVariables/DataStore_Mails.lua
@@ -8,8 +8,8 @@ DataStore_MailsDB = {
 	["global"] = {
 		["Characters"] = {
 			["Default.鱼人摩洛克.特变电工"] = {
-				["lastUpdate"] = 1650374097,
-				["lastVisitDate"] = "2022/04/19 21:14",
+				["lastUpdate"] = 1650706394,
+				["lastVisitDate"] = "2022/04/23 17:33",
 			},
 			["Default.鱼人摩洛克.Javascript"] = {
 				["lastUpdate"] = 1645540439,