+++ WTF/Account/197404344#2/SavedVariables/DataStore_Spells.lua
@@ -8,7 +8,7 @@ DataStore_SpellsDB = {
 	["global"] = {
 		["Characters"] = {
 			["Default.鱼人摩洛克.特变电工"] = {
-				["lastUpdate"] = 1650383355,
+				["lastUpdate"] = 1650708008,
 				["Spells"] = {
 					["武器"] = {
 						"11578|等级 3", -- [1]