+++ WTF/Account/197404344#2/SavedVariables/MeetingHorn.lua
@@ -22,22 +22,6 @@ MEETINGHORN_DB = {
 					["scale"] = 1,
 				},
 			},
-			["goodleader"] = {
-				["cache"] = {
-					["祖尔格拉布"] = {
-						["leader"] = "大佬来了-鱼人摩洛克",
-						["timestamp"] = 1650298403,
-						["leaderGuid"] = "Player-5458-0359779F",
-						["raidName"] = "祖尔格拉布",
-					},
-					["奥妮克希亚的巢穴"] = {
-						["leaderGuid"] = "Player-5458-03538A5B",
-						["timestamp"] = 1650382842,
-						["leader"] = "楠神-鱼人摩洛克",
-						["raidName"] = "奥妮克希亚的巢穴",
-					},
-				},
-			},
 		},
 		["Javascript - 鱼人摩洛克"] = {
 		},