+++ WTF/Account/197404344#2/鱼人摩洛克/特变电工/SavedVariables/TradeLogUnlimited.lua
@@ -16,9 +16,233 @@ TradeLogUnlimited_TradesHistory = {
 		},
 		["who"] = "菲奧娜",
 		["when"] = "2022-04-19 23:50:54",
+		["result"] = "complete",
+		["where"] = "铁炉堡",
+	}, -- [1]
+	{
+		["targetMoney"] = 0,
+		["playerMoney"] = 10000000,
+		["id"] = 2,
+		["playerItems"] = {
+		},
+		["targetItems"] = {
+			{
+				["name"] = "泰坦护腿",
+				["numItems"] = 1,
+				["tCount"] = 1,
+				["isUsable"] = true,
+				["itemLink"] = "|cffa335ee|Hitem:22385::::::::60:::::::::|h[泰坦护腿]|h|r",
+				["texture"] = 134584,
+			}, -- [1]
+		},
+		["player"] = "特变电工",
+		["events"] = {
+			"TRADE_SHOW", -- [1]
+			"TRADE_CLOSED", -- [2]
+			"TRADE_CLOSED", -- [3]
+		},
+		["who"] = "潮汐海灵",
+		["when"] = "2022-04-23 12:13:26",
+		["result"] = "complete",
+		["where"] = "铁炉堡",
+	}, -- [2]
+	{
+		["targetMoney"] = 0,
+		["playerMoney"] = 800000,
+		["id"] = 3,
+		["playerItems"] = {
+			[7] = {
+				["name"] = "泰坦护腿",
+				["numItems"] = 1,
+				["enchantment"] = "次级奥术融合",
+				["itemLink"] = "|cffa335ee|Hitem:22385::::::::60:::::::::|h[泰坦护腿]|h|r",
+				["texture"] = 134584,
+			},
+		},
+		["targetItems"] = {
+		},
+		["player"] = "特变电工",
+		["events"] = {
+			"TRADE_SHOW", -- [1]
+			"TRADE_CLOSED", -- [2]
+			"TRADE_CLOSED", -- [3]
+		},
+		["who"] = "附魔小店",
+		["when"] = "2022-04-23 12:21:41",
+		["result"] = "complete",
+		["where"] = "铁炉堡",
+	}, -- [3]
+	{
+		["targetMoney"] = 0,
+		["playerMoney"] = 300000,
+		["id"] = 4,
+		["playerItems"] = {
+			[7] = {
+				["name"] = "祭祀护手",
+				["numItems"] = 1,
+				["enchantment"] = "附魔手套 - 强效力量",
+				["itemLink"] = "|cff0070dd|Hitem:22714::::::::60:::::::::|h[祭祀护手]|h|r",
+				["texture"] = 132965,
+			},
+		},
+		["reason"] = "self",
+		["targetItems"] = {
+		},
+		["player"] = "特变电工",
+		["where"] = "铁炉堡",
+		["who"] = "二狗不怕疯",
+		["when"] = "2022-04-23 12:26:03",
+		["result"] = "cancelled",
+	}, -- [4]
+	{
+		["targetMoney"] = 0,
+		["playerMoney"] = 300000,
+		["id"] = 5,
+		["playerItems"] = {
+			[7] = {
+				["texture"] = 132965,
+				["itemLink"] = "|cff0070dd|Hitem:22714::::::::60:::::::::|h[祭祀护手]|h|r",
+				["name"] = "祭祀护手",
+				["numItems"] = 1,
+			},
+		},
+		["reason"] = "self",
+		["targetItems"] = {
+		},
+		["player"] = "特变电工",
+		["where"] = "铁炉堡",
+		["who"] = "二狗不怕疯",
+		["when"] = "2022-04-23 12:26:15",
+		["result"] = "cancelled",
+	}, -- [5]
+	{
+		["targetMoney"] = 0,
+		["playerMoney"] = 300000,
+		["id"] = 6,
+		["playerItems"] = {
+			[7] = {
+				["name"] = "祭祀护手",
+				["numItems"] = 1,
+				["enchantment"] = "附魔手套 - 强效力量",
+				["itemLink"] = "|cff0070dd|Hitem:22714::::::::60:::::::::|h[祭祀护手]|h|r",
+				["texture"] = 132965,
+			},
+		},
+		["targetItems"] = {
+		},
+		["player"] = "特变电工",
+		["events"] = {
+			"TRADE_SHOW", -- [1]
+			"TRADE_CLOSED", -- [2]
+			"TRADE_CLOSED", -- [3]
+		},
+		["who"] = "二狗不怕疯",
+		["when"] = "2022-04-23 12:26:49",
+		["result"] = "complete",
+		["where"] = "铁炉堡",
+	}, -- [6]
+	{
+		["targetMoney"] = 0,
+		["playerMoney"] = 10000,
+		["id"] = 7,
+		["playerItems"] = {
+		},
+		["targetItems"] = {
+		},
+		["player"] = "特变电工",
+		["events"] = {
+			"TRADE_SHOW", -- [1]
+			"TRADE_CLOSED", -- [2]
+			"TRADE_CLOSED", -- [3]
+		},
+		["who"] = "联盟小红手",
+		["when"] = "2022-04-23 12:30:22",
+		["result"] = "complete",
+		["where"] = "铁炉堡",
+	}, -- [7]
+	{
+		["targetMoney"] = 0,
+		["playerMoney"] = 10000,
+		["id"] = 8,
+		["playerItems"] = {
+		},
+		["targetItems"] = {
+			{
+				["itemLink"] = "|cffffffff|Hitem:22895::::::::60:::::::::|h[魔法肉桂面包]|h|r",
+				["name"] = "魔法肉桂面包",
+				["numItems"] = 20,
+				["tCount"] = 120,
+				["isUsable"] = true,
+				["texture"] = 134029,
+			}, -- [1]
+			{
+				["name"] = "魔法肉桂面包",
+				["numItems"] = 20,
+				["itemLink"] = "|cffffffff|Hitem:22895::::::::60:::::::::|h[魔法肉桂面包]|h|r",
+				["isUsable"] = true,
+				["texture"] = 134029,
+			}, -- [2]
+			{
+				["name"] = "魔法肉桂面包",
+				["numItems"] = 20,
+				["itemLink"] = "|cffffffff|Hitem:22895::::::::60:::::::::|h[魔法肉桂面包]|h|r",
+				["isUsable"] = true,
+				["texture"] = 134029,
+			}, -- [3]
+			{
+				["name"] = "魔法肉桂面包",
+				["numItems"] = 20,
+				["itemLink"] = "|cffffffff|Hitem:22895::::::::60:::::::::|h[魔法肉桂面包]|h|r",
+				["isUsable"] = true,
+				["texture"] = 134029,
+			}, -- [4]
+			{
+				["name"] = "魔法肉桂面包",
+				["numItems"] = 20,
+				["itemLink"] = "|cffffffff|Hitem:22895::::::::60:::::::::|h[魔法肉桂面包]|h|r",
+				["isUsable"] = true,
+				["texture"] = 134029,
+			}, -- [5]
+			{
+				["name"] = "魔法肉桂面包",
+				["numItems"] = 20,
+				["itemLink"] = "|cffffffff|Hitem:22895::::::::60:::::::::|h[魔法肉桂面包]|h|r",
+				["isUsable"] = true,
+				["texture"] = 134029,
+			}, -- [6]
+		},
+		["player"] = "特变电工",
+		["events"] = {
+			"TRADE_SHOW", -- [1]
+			"TRADE_CLOSED", -- [2]
+			"TRADE_CLOSED", -- [3]
+		},
+		["who"] = "联盟小红手",
+		["when"] = "2022-04-23 17:37:14",
 		["where"] = "铁炉堡",
 		["result"] = "complete",
-	}, -- [1]
+	}, -- [8]
+	{
+		["targetMoney"] = 0,
+		["playerMoney"] = 50000,
+		["id"] = 9,
+		["playerItems"] = {
+		},
+		["targetItems"] = {
+		},
+		["player"] = "特变电工",
+		["events"] = {
+			"TRADE_CLOSED", -- [1]
+			"TRADE_CLOSED", -- [2]
+			"TRADE_SHOW", -- [3]
+			"TRADE_CLOSED", -- [4]
+			"TRADE_CLOSED", -- [5]
+		},
+		["who"] = "糖玉米",
+		["when"] = "2022-04-23 17:39:27",
+		["where"] = "希利苏斯",
+		["result"] = "complete",
+	}, -- [9]
 }
-TradeLogUnlimited_Announce_Checked = nil
+TradeLogUnlimited_Announce_Checked = false
 TradeLogUnlimited_AnnounceChannel = "WHISPER"