+++ Interface/AddOns/honorspy/CHANGELOG.md
@@ -1,6 +1,18 @@
 # HonorSpy
 
-## [1.9.1](https://github.com/kakysha/HonorSpy/tree/1.9.1) (2022-04-08)
-[Full Changelog](https://github.com/kakysha/HonorSpy/compare/1.9.0...1.9.1) [Previous Releases](https://github.com/kakysha/HonorSpy/releases)
+## [1.9.2](https://github.com/kakysha/HonorSpy/tree/1.9.2) (2022-04-19)
+[Full Changelog](https://github.com/kakysha/HonorSpy/compare/1.9.1...1.9.2) [Previous Releases](https://github.com/kakysha/HonorSpy/releases)
 
-- Wago support  
+- 1.9.2  
+- Merge pull request #216 from Slivo-fr/bugfix/ignore-failed-reset  
+    Rejects players with incorrect weekly reset data for the first hours after weekly reset  
+- Rejects faulty reset broadcast for the first 12 hours of the weekly reset  
+- Merge pull request #215 from Slivo-fr/bugfix/sort-csv-estimated-honor  
+    Use estimated week honor for csv export sorting  
+- Use estimated week honor for csv export sorting  
+- Merge pull request #214 from Slivo-fr/bugfix/missing-csv-fields  
+    Use empty string instead of nil for estHonor in standing table  
+- Merge pull request #217 from Slivo-fr/patch-4  
+    Makes player with 0 honor (but 15hk) valid  
+- Makes player with 0 honor (but 15hk) valid  
+- Use empty string instead of nil for estHonor in standing table